Install scripts updated

+++ install/nvm.sh
@@ -1,15 +1,17 @@
 #!/bin/bash
 
-curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.34.0/install.sh | bash
-cd ~/.nvm
-. ./nvm.sh
-nvm install node
-nvm use node
+if [  -f "~/.nvm/nvm.sh" ]; then
+  curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.34.0/install.sh | bash
+  cd ~/.nvm
+  . ./nvm.sh
+  nvm install --lts
+  nvm use node
+  nvm alias default node
+fi
 
 packages=(
     http-server
     nodemon
-    gulp
     typescript
     tldr
     gitmoji-cli
